Remove old scripting bindings
